Fixed some warnings.

+++ Core/include/SmartHistogram.h
@@ -181,18 +181,18 @@ public:
 template<>
 class SmartHistogram<TH1D> : public TH1D, public AbstractHistogram {
 public:
-    SmartHistogram(const std::string& name, size_t nbins, double low, double high)
-        : TH1D(name.c_str(), name.c_str(), static_cast<int>(nbins), low, high), AbstractHistogram(name), store(true),
+    SmartHistogram(const std::string& name, int nbins, double low, double high)
+        : TH1D(name.c_str(), name.c_str(), nbins, low, high), AbstractHistogram(name), store(true),
           use_log_y(false), max_y_sf(1), divide_by_bin_width(false) {}
 
     SmartHistogram(const std::string& name, const std::vector<double>& bins)
         : TH1D(name.c_str(), name.c_str(), static_cast<int>(bins.size()) - 1, bins.data()), AbstractHistogram(name),
           store(true), use_log_y(false), max_y_sf(1), divide_by_bin_width(false) {}
 
-    SmartHistogram(const std::string& name, size_t nbins, double low, double high, const std::string& x_axis_title,
+    SmartHistogram(const std::string& name, int nbins, double low, double high, const std::string& x_axis_title,
                    const std::string& y_axis_title, bool _use_log_y, double _max_y_sf, bool _divide_by_bin_width,
                    bool _store)
-        : TH1D(name.c_str(), name.c_str(), static_cast<int>(nbins), low, high), AbstractHistogram(name), store(_store),
+        : TH1D(name.c_str(), name.c_str(), nbins, low, high), AbstractHistogram(name), store(_store),
           use_log_y(_use_log_y), max_y_sf(_max_y_sf), divide_by_bin_width(_divide_by_bin_width)
     {
         SetXTitle(x_axis_title.c_str());
@@ -260,15 +260,15 @@ template<>
 class SmartHistogram<TH2D> : public TH2D, public AbstractHistogram {
 public:
     SmartHistogram(const std::string& name,
-                   size_t nbinsx, double xlow, double xup,
-                   size_t nbinsy, double ylow, double yup)
-        : TH2D(name.c_str(), name.c_str(), static_cast<int>(nbinsx), xlow, xup, static_cast<int>(nbinsy), ylow, yup),
+                   int nbinsx, double xlow, double xup,
+                   int nbinsy, double ylow, double yup)
+        : TH2D(name.c_str(), name.c_str(), nbinsx, xlow, xup, nbinsy, ylow, yup),
           AbstractHistogram(name), store(true), use_log_y(false), max_y_sf(1) {}
 
-    SmartHistogram(const std::string& name, size_t nbinsx, double xlow, double xup, size_t nbinsy, double ylow,
+    SmartHistogram(const std::string& name, int nbinsx, double xlow, double xup, int nbinsy, double ylow,
                    double yup, const std::string& x_axis_title, const std::string& y_axis_title, bool _use_log_y,
                    double _max_y_sf, bool _store)
-        : TH2D(name.c_str(), name.c_str(), static_cast<int>(nbinsx), xlow, xup, static_cast<int>(nbinsy), ylow, yup),
+        : TH2D(name.c_str(), name.c_str(), nbinsx, xlow, xup, nbinsy, ylow, yup),
           AbstractHistogram(name), store(_store), use_log_y(_use_log_y), max_y_sf(_max_y_sf)
     {
         SetXTitle(x_axis_title.c_str());
@@ -363,7 +363,7 @@ template<>
 struct HistogramFactory<TH1D> {
 private:
     struct HistogramParameters {
-        size_t nbins;
+        int nbins;
         double low;
         double high;
     };
